Day40-FlightFinder-Part2
Edited the code to send emails to parties that sign up using the replit program.
 - This code will only be on replit.

This entire part was extremely time consuming for little pay off.

Small chunk of the code doesn't work properly due to google's flight website not accepting parameters that were given in one of the course documents. Most specifically the URL link that it's supposed to send you, so that you may immediately book the flight through google flights website. Commented this code out, everything else works as intended.