Day 47 - Amazon "Automated" Price Checker
small program that checks price of an item listed on amazon.

If discount is above specified price, sends email alert.

As usual, all constants have been stripped and will need to be populated. All have been marked with a @TODO tag for easy finding.

Only works for a single item but could easily be changed to accommodate multiple items.