Day 49 - LinkedIn Job Application Automation
Small program that signs in to linkedin, and then applies to all job listings on the first page of the search params that are passed in through the url.

Currently does not actually apply. There is functionality for it that's commented out, currently will open the "Easy Apply" option and fill in the phone number entry, but then immediately close it and continue on. All other entries that don't have the "Easy Apply" option are saved.