Day 52 - Data Entry Apartment Automation
Small program that grabs information from Zillow, based on a certain URL that is passed in.

Then takes the info grabbed, and fills out a google form, which can then be transformed into a google sheet. Second part cannot be automated. Due to google labeling as an insecure browser, cannot automate the sign in process to convert to sheet.