Day 53-59
Includes the Tindog website and Blog website that we created while learning HTML CSS and Jinja.

Tindog has little to no functionality. A lot of HTML was already set as base. This project was mostly to learn CSS.

The Blog website was a template website that was downloaded from: https://startbootstrap.com/previews/clean-blog
All that was done here from me was adding the Jinja to account for the blogs and blog previews on the main page. Also connecting the pages for basic navigation inside the website using Flask.