Day 69 Update 2
Added most functionality to the site.

Made a redirect when trying to access content that needs user to be signed in, sends to login page, and then returns to original destination if possible.

Security change that limits the post editing and deletion to the person who originally posted the post.
 - Will/should be adding a bypass for admins, so they can freely edit others' posts and delete if needed.