Day 69 Update 4
Just about everything is set up, all that is left is creating the logic behind letting users comment on blog posts.

Set up a relationship between user and blog post so that in the future if we enabled more people to post, it'd be easy to keep track of who posted what.

A few other tweaks not worth mentioning.